+++ src/App.js
@@ -8,7 +8,7 @@ import TechniciansMain from './components/Technicians/TechniciansMain';
 import BoilerTypesApp from './components/boilerTypes/BoilerTypesApp';
 import Layout from './components/Layout';
 import Appointments from './components/Appointments';
-import BoilersApp from './components/Boilers';
+import Boilers from './components/Boilers';
 
 function App() {
   return (
@@ -17,7 +17,7 @@ function App() {
         <Switch>
           <Route path="/buildings" component={BuildingApp} />
           <Route path="/companies" component={Companies} />
-          <Route path="/boilers" component={BoilersApp} />
+          <Route path="/boilers" component={Boilers} />
           <Route path="/boilerTypes" component={BoilerTypesApp} />
           <Route path="/technicians" component={TechniciansMain} />
           <Route path="/appointments" component={Appointments} />
